refactor: keep combined provider settings within the source file limit
The combined Copilot and oFOX settings pane exceeded the enforced 500-line limit. Extract only Copilot rendering and reasoning options, passing existing fields and capability override callbacks unchanged.

Validation: TypeScript, lint, and verify-gate-coverage passed; all 576 verification files remain reachable from npm test.
